Upgrade default box image to trusty
I left the default box as swift-all-in-one, so whatever you have in your
`vagrant box list` as swift-all-in-one will still get used for new boxes if
you destory/up.  But if you change the localrc VAGRANT_BOX to trusty/precise
it will download the appropriate box image and use that (you can flip back and
forth but you have to destroy the running vm first).

If you set VAGRANT_BOX to something besdies one of the named boxes available
at the top of the Vagrantfile you'll have to make sure the box image is
already downloaded and added to vagrant's box list.

It's unknown how any of this works with Vagrant ~1.7 with the new
config.vm.name support for Atlas.  I'm still running 1.4.3.