Change default Keystone branch
This vsaio works with master branch of Keystone so far but Keystone
changed the spec several times during Mitaka-Newton time period so
this patch changes the default Keystone branch to stable/liberty
which I confirmed behavior with this vsaio.

Note: If you specify master branch of Keystone in localrc with current
implementation, this vsaio doesn't work.