Remove hardlink hack from vtox
Someone (pip?  distutils?) used to assume the install source and install
target were on the same filesystem for doing a development install.
That doesn't seem to be the case with later-ish tools and it was causing
problems with tox making it think it needed to reinstall the venv
everytime because the mtime on the setup.py file was > than the mtime on
the swift.egg-info directory.

I think the tox behavior is probably a bug - but removing swift.egg-info
and letting reinstall create it seems get things going - leaving only
the "not with the always modifying the setup.py" to make it property
skip the venv reinstall and ~80s of setup before running any tests OMM