Adding .aws/credentials file
Newer versions of the aws client will look for credentials
in a separate ~/.aws/credentials file and ignore those
in ~/.aws/config, causing authentication to fail.

This change adds the credentials to the ~/.aws/credentials file also,
so that any users with a newer aws client will find the creds,
while older clients will continue to find them in ~/.aws/config.