Add ability to configure a statsd_exporter/prometheus metrics pipeline
Configure statsd_exporter instances for each of

- proxy-server
- backend nodes 1 through 4
- no-auth proxy

These will expose /metrics endpoints on ports 9100-9105 which developers
can check directly.

Configure prometheus to scrape these endpoints every 10s and retain data
for up to a day; developers can then create ad-hoc graphs at

   http://saio:9090/graph

Note that to enable proper per-service attribution of metrics, I needed
to update a bunch of log_name config options; for consistency's sake,
this is enabled even when STATSD_EXPORTER=false (which is the default).

In the future we may want to install grafana and build out some
dashboards so we can flip between pre-labeled metrics, labeled metrics,
and legacy-mode in a post-labeled-metrics world and verify we can do
that with no gaps in graphs.