fix clean deployment
Running vagrant up was failing when there was no existing VM.
The "fix certifi" task in `source.rb` requires the certificate, which is
not created until the "create cert" task in `config.rb` later in the process.

This change ensures that the VM creates successfully by making the
"fix certifi" task in `source.rb` conditional on the certificate existing.

It also restores the "fix certifi" task to config.rb so that it is run both
on provision, and with the `reinstallswift` script.