Add ring_version option to Vsaio
By default swift-ring-builder and remakerings will create rings in
format_version/serialization version 1. But we want to have the ability
to choose which version to run.

This patch adds a new option ring_version or in localrc RING_VERSION.
This defaults to 1. But when set to 2 will build rings in v2.

It does this by simply adding --format-version=<ring_version> to the
rings recipe which is called be remakerings in vSAIO.